# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/create": {
            "post": {
                "description": "Создание новой персоны с обогащением данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "people"
                ],
                "summary": "Создать персону",
                "parameters": [
                    {
                        "description": "Данные для создания персоны",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/em-task_internal_repository.CreatePersonDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное создание",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.PersonResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/delete/{id}": {
            "delete": {
                "description": "Удаление персоны по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "people"
                ],
                "summary": "Удалить персону",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "ID персоны",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное удаление",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.SuccessfullyDeletedResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/edit/{id}": {
            "patch": {
                "description": "Редактирование данных персоны по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "people"
                ],
                "summary": "Редактировать персону",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "ID персоны",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновленные данные",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/em-task_internal_repository.FullPersonDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленная персона",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.PersonResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/get": {
            "get": {
                "description": "Получение списка персон с использованием параметров фильтрации, лимита и смещения",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "people"
                ],
                "summary": "Получить список персон",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 10,
                        "description": "Лимит",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 0,
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 30,
                        "description": "Возраст",
                        "name": "age",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 18,
                        "description": "Возраст от",
                        "name": "age_from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 40,
                        "description": "Возраст до",
                        "name": "age_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "male",
                        "description": "Пол",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "John",
                        "description": "Имя",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "US",
                        "description": "Национальность",
                        "name": "nationality",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "William",
                        "description": "Отчество",
                        "name": "patronymic",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "Smith",
                        "description": "Фамилия",
                        "name": "surname",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.PersonResponse"
                        }
                    },
                    "400": {
                        "description": "Неверные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.InternalServerErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "em-task_internal_repository.CreatePersonDTO": {
            "type": "object",
            "required": [
                "name",
                "surname"
            ],
            "properties": {
                "name": {
                    "description": "Имя",
                    "type": "string",
                    "example": "John"
                },
                "patronymic": {
                    "description": "Отчество",
                    "type": "string",
                    "example": "William"
                },
                "surname": {
                    "description": "Фамилия",
                    "type": "string",
                    "example": "Smith"
                }
            }
        },
        "em-task_internal_repository.FullPersonDTO": {
            "type": "object",
            "required": [
                "name",
                "surname"
            ],
            "properties": {
                "age": {
                    "description": "Возраст",
                    "type": "integer",
                    "minimum": 0,
                    "example": 30
                },
                "gender": {
                    "description": "Пол",
                    "type": "string",
                    "example": "male"
                },
                "id": {
                    "description": "ID",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "Имя",
                    "type": "string",
                    "example": "John"
                },
                "nationality": {
                    "description": "Национальность",
                    "type": "string",
                    "example": "US"
                },
                "patronymic": {
                    "description": "Отчество",
                    "type": "string",
                    "example": "William"
                },
                "surname": {
                    "description": "Фамилия",
                    "type": "string",
                    "example": "Smith"
                }
            }
        },
        "internal_handlers.BadRequestResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "Invalid ID format"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "internal_handlers.InternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "message": {
                    "type": "string",
                    "example": "Some error message"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "internal_handlers.PersonResponse": {
            "type": "object",
            "properties": {
                "Age": {
                    "type": "integer",
                    "example": 49
                },
                "CreatedAt": {
                    "type": "string",
                    "example": "2025-05-01T03:35:20.374096098+03:00"
                },
                "DeletedAt": {
                    "type": "string",
                    "example": ""
                },
                "Gender": {
                    "type": "string",
                    "example": "female"
                },
                "ID": {
                    "type": "integer",
                    "example": 42
                },
                "Name": {
                    "type": "string",
                    "example": "John"
                },
                "Nationality": {
                    "type": "string",
                    "example": "SK"
                },
                "Patronymic": {
                    "type": "string",
                    "example": ""
                },
                "Surname": {
                    "type": "string",
                    "example": "Smith"
                },
                "UpdatedAt": {
                    "type": "string",
                    "example": "2025-05-01T03:35:20.374096098+03:00"
                }
            }
        },
        "internal_handlers.SuccessfullyDeletedResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "Person successfully deleted"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        }
    }
}